tests/wpt/web-platform-tests/offscreen-canvas/the-canvas-state/2d.state.saverestore.underflow.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.state.saverestore.underflow</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.state.saverestore.underflow</h1>
<p class="desc">restore() with an empty stack has no effect</p>


<script>
var t = async_test("restore() with an empty stack has no effect");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

for (var i = 0; i < 16; ++i)
    ctx.restore();
ctx.lineWidth = 0.5;
ctx.restore();
_assertSame(ctx.lineWidth, 0.5, "ctx.lineWidth", "0.5");

t.done();

});
</script>
